Ecom_app Flutter/Dart, fifteenth Commit. Created and added apis that connect and 'speak' with the server to show data to the frontend. Frontend backend communication and connection made.

+++ lib/widgets/base_text_widget.dart
@@ -5,7 +5,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../entities/values/colors.dart';
 
-Widget reusableText(String text, {Color color=AppColors.primaryText, int fontSize = 16, FontWeight fontWeight=FontWeight.bold}){
+Widget reusableText(String text, {Color color=AppColors.primaryText, double fontSize = 16, FontWeight fontWeight=FontWeight.bold}){
   return Text(
     text,
     style: TextStyle(
